OpenSSLX509Certificate: mark mContext as transient
Since mContext should not participate in the serialization process,
hide it with the transient qualifier. This will prevent the field from
initialization during the unserialization of this class. Then of course
the instance will be in a valid state.

(cherry picked from commit 8d57b9dbbd883422a0ff02083bfcf637b097e504)

CYNGNOS-446
Bug: 21437603
Change-Id: Ie9453c16d11820a91caff92c3f7b326d12f8a8f4